refactor: 닉네임 관련 리팩터링
관련된 클래스끼리 패키지로 분리
가독성을 위한 메소드 위치 이동
불필요한 메서드 분리 제거
