+++ README.md
@@ -36,7 +36,7 @@
 - 닉네임
   - [x] 닉네임은 null일 수 없다.
   - [x] 닉네임 앞뒤에 white space가 있으면 white space는 제거된다.
-  - [ ] 닉네임은 공백일 수 없다.
+  - [x] 닉네임은 공백일 수 없다.
   - [x] 닉네임은 5자를 초과할 수 없다.
 
 ### 프로그래밍 요구사항
