+++ src/main/java/racingcar/nickname/CanNotBlankNicknameException.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.nickname;
 
 public class CanNotBlankNicknameException extends RuntimeException {
