+++ src/main/java/racingcar/nickname/CanNotOverFiveCharacterNicknameException.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.nickname;
 
 public class CanNotOverFiveCharacterNicknameException extends RuntimeException {
+
     private static final String MESSAGE = "다섯자를 초과하는 닉네임은 사용할 수 없습니다.";
 
     public CanNotOverFiveCharacterNicknameException() {
